+++ packages/core/src/source/types.ts
@@ -29,16 +29,3 @@ export type InferMetaType<Utils extends LoaderOutput<any>> =
  * @internal
  */
 export type UrlFn = (slugs: string[], locale?: string) => string;
-
-/**
- * @internal
- */
-export interface FileData {
-  meta: {
-    data: MetaData;
-  };
-  file: {
-    slugs: string[];
-    data: PageData;
-  };
-}