+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # next-docs-ui
 
+## 13.0.4
+
+### Patch Changes
+
+- Updated dependencies [5355391]
+  - fumadocs-core@13.0.4
+
+## 13.0.3
+
+### Patch Changes
+
+- Updated dependencies [978342f]
+  - fumadocs-core@13.0.3
+
 ## 13.0.2
 
 ### Patch Changes